refactor: Move install logic to dedicated module
`mod_management` module didn't exist when this PR was created
