+++ src-tauri/src/main.rs
@@ -11,10 +11,9 @@ use std::{
 
 use app::{
     check_is_flightcore_outdated, check_is_valid_game_path, check_northstar_running,
-    check_origin_running, convert_release_candidate_number, fc_download_mod_and_install,
-    find_game_install_location, get_enabled_mods, get_host_os, get_log_list,
-    get_northstar_version_number, install_northstar, launch_northstar, linux_checks_librs,
-    GameInstall, NorthstarMod,
+    check_origin_running, convert_release_candidate_number, find_game_install_location,
+    get_enabled_mods, get_host_os, get_log_list, get_northstar_version_number, install_northstar,
+    launch_northstar, linux_checks_librs, GameInstall, NorthstarMod,
 };
 
 mod github;
@@ -24,7 +23,9 @@ mod repair_and_verify;
 use repair_and_verify::{disable_all_but_core, verify_game_files};
 
 mod mod_management;
-use mod_management::{get_installed_mods_and_properties, set_mod_enabled_status};
+use mod_management::{
+    fc_download_mod_and_install, get_installed_mods_and_properties, set_mod_enabled_status,
+};
 
 use tauri::Manager;
 use tauri_plugin_store::PluginBuilder;
